requirements.txt
ipython
django-extensions
